requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
Django==5.1.4
Flask==3.1.0
gevent==24.11.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.0.0
python-dotenv==1.0.1
sorl-thumbnail==12.11.0
sqlparse==0.5.3
Werkzeug==3.1.3
wsgi-oauth2==0.2.2
zope.event==5.0
zope.interface==7.2
